demos/recipenet/requirements.txt
torch==2.3.0+cu121
torchvision==0.18.0+cu121
Pillow
gradio==4.36.0
